Fixed floating window focus bug, sort of
There's an issue with the selected instance not matching up with the editor window. Although the selected instance is the only element with keyboard focus, it has some weird behavior. I would have to select the instance in the explorer, and the overlay would conform to that instance's open script editor.

I should probably see if there are other sources for the name of the script other than the element with keyboard focus.