Changed how the script name was retrieved
I suppose since we have to open the script to run it, we can store the name of the script there and use it to find our script editor window.

So this marks everything I wanted to solve with this program. I might refactor this on another pull reguest, I'm not sure.